requirements.txt
requests
tweepy
python-dotenv
boto3